% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_move.R
\name{black_move}
\alias{black_move}
\title{Makes a move for black}
\usage{
black_move(move, startingboard)
}
\arguments{
\item{move}{the move notation in PGN format}

\item{startingboard}{the layout of the board before the move}
}
\description{
Makes a move for black based on PGN format and a starting chess board.
}
