% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_games_id.R
\name{download_games_id}
\alias{download_games_id}
\title{Download chess games by ID}
\usage{
download_games_id(idvector, destination)
}
\arguments{
\item{idvector}{a vector of game IDs}

\item{destination}{a location in which to save the games}
}
\description{
Allows users to download specific games by ID on chessgames.com
}
