% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_random_games.R
\name{download_random_games}
\alias{download_random_games}
\title{Downloads random chess games}
\usage{
download_random_games(number, destination)
}
\arguments{
\item{number}{the number of games to download}

\item{destination}{an empty directory in which to save the games}
}
\description{
Downloads random chess games in PGN format from chessgames.com
}
\details{
The destination specified should be an EMPTY DIRECTORY. The function counts the files in the directory to determine how many more random games to download.

For example, if you specify 100 random games and your directory already contains 50 files, then the function will stop after only 50 games.

The destination should end in a forward slash -- for example, "data/"
}
