% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boards.R
\name{get_boards}
\alias{get_boards}
\title{Produces a chess game for analysis}
\usage{
get_boards(moves)
}
\arguments{
\item{moves}{the outputs of the get_moves() function}
}
\description{
Accepts moves from the get_moves() function and outputs a chess game object.
}
