% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moves.R
\name{get_moves}
\alias{get_moves}
\title{Gets individual moves in PGN}
\usage{
get_moves(text)
}
\arguments{
\item{text}{the text from a PGN format game}
}
\description{
Cleans PGN games and produces a list of moves
}
