% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_move.R
\name{white_move}
\alias{white_move}
\title{Makes a move for white}
\usage{
white_move(move, startingboard)
}
\arguments{
\item{move}{the move notation in PGN format}

\item{startingboard}{the layout of the board before the move}
}
\description{
Makes a move for white based on PGN format and a starting chess board.
}
